moved things to dataAccess.js from dataAccessChat.js
